@@ -10,17 +10,18 @@
   },
   "dependencies": {
     "autoprefixer": "10.4.14",
-    "bcrypt": "^5.1.0",
-    "mongodb": "^5.3.0",
-    "mongoose": "^7.1.0",
-    "next": "13.3.4",
-    "next-auth": "^4.22.1",
-    "react": "18.2.0",
-    "react-dom": "18.2.0"
-   
+    "bcrypt": "^5.1.1",
+    "mongodb": "^6.12.0",
+    "mongoose": "^8.9.2",
+    "next": "15.1.2",
+    "next-auth": "^4.24.11",
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0"
   },
   "devDependencies": {
     "postcss": "^8",
-    "tailwindcss": "^3.4.1"
+    "tailwindcss": "^3.4.1",
+    "webpack": "^5.98.0",
+    "webpack-cli": "^6.0.1"
   }
 }
